+++ app/src/main/java/org/ole/planet/myplanet/ui/viewer/PDFReaderActivity.kt
@@ -37,14 +37,14 @@ class PDFReaderActivity : AppCompatActivity(), OnPageChangeListener, OnLoadCompl
         activityPdfReaderBinding = ActivityPdfreaderBinding.inflate(layoutInflater)
         setContentView(activityPdfReaderBinding.root)
         audioRecorderService = AudioRecorderService().setAudioRecordListener(this)
-        recordingPermissionManager = AudioRecorderPermission(this, this, audioRecorderService)
+        audioRecorderService.setCaller(this, this)
         mRealm = DatabaseService(this).realmInstance
         if (intent.hasExtra("resourceId")) {
             val resourceID = intent.getStringExtra("resourceId")
             library = mRealm.where(RealmMyLibrary::class.java).equalTo("id", resourceID).findFirst()!!
         }
         renderPdfFile()
-        activityPdfReaderBinding.fabRecord.setOnClickListener { recordingPermissionManager.onRecordClicked()}
+        activityPdfReaderBinding.fabRecord.setOnClickListener { audioRecorderService.onRecordClicked()}
         activityPdfReaderBinding.fabPlay.setOnClickListener {
             if (this::library.isInitialized && !TextUtils.isEmpty(library.translationAudioPath)) {
                 openAudioFile(this, library.translationAudioPath)
